Move logging to header
